Check in list fixes

+++ backend/app/Services/Application/Handlers/CheckInList/DeleteCheckInListHandler.php
@@ -2,6 +2,7 @@
 
 namespace HiEvents\Services\Application\Handlers\CheckInList;
 
+use HiEvents\Exceptions\ResourceConflictException;
 use HiEvents\Repository\Interfaces\CheckInListRepositoryInterface;
 use Symfony\Component\Routing\Exception\ResourceNotFoundException;
 
@@ -23,6 +24,12 @@ class DeleteCheckInListHandler
 
         if ($checkInList === null) {
             throw new ResourceNotFoundException(__('Check-in list not found'));
+        }
+
+        if ($checkInList->getIsSystemDefault()) {
+            throw new ResourceConflictException(
+                __('The default check-in list can\'t be deleted.')
+            );
         }
 
         $this->checkInListRepository->deleteWhere([
